Moved more components into reusables folder
